imageserver/requirements.txt
Flask==1.0.2
Pillow==5.4.1
gunicorn[gthread]==19.9.0
